+++ server/routes/discRoutes.js
@@ -5,6 +5,7 @@ const discController = require('../controllers/discController');
 router.get('/questions/:no', discController.getQuestionsByNo);
 router.get('/result/:userId', discController.getResultsById);
 router.get('/master-questions', discController.getMasterQuestions);
+router.post('/change-password', discController.submitChangePassword);
 router.post('/submit-answer', discController.submitAnswer);
 
 module.exports = router;